change: examples and projects
