+++ .storybook/index.css
@@ -2500,10 +2500,6 @@ select {
   max-width: 320px;
 }
 
-.max-w-\[35\%\] {
-  max-width: 35%;
-}
-
 .max-w-\[40\%\] {
   max-width: 40%;
 }
@@ -2546,10 +2542,6 @@ select {
 
 .max-w-xs {
   max-width: 20rem;
-}
-
-.max-w-\[30\%\] {
-  max-width: 30%;
 }
 
 .flex-1 {
@@ -7917,14 +7909,6 @@ StackedList for ToC
   }
 }
 
-@container (min-width: 24rem) {
-  @media (min-width: 640px) {
-    .\@sm\:sm\:max-w-\[25\%\] {
-      max-width: 25%;
-    }
-  }
-}
-
 @media (min-width: 768px) {
   .md\:float-right {
     float: right;
@@ -8351,10 +8335,6 @@ StackedList for ToC
       width: 1.5rem;
     }
 
-    .\@sm\:md\:max-w-\[25\%\] {
-      max-width: 25%;
-    }
-
     .\@sm\:md\:space-y-2 > :not([hidden]) ~ :not([hidden]) {
       --tw-space-y-reverse: 0;
       margin-top: calc(0.5rem * calc(1 - var(--tw-space-y-reverse)));
@@ -8417,18 +8397,6 @@ StackedList for ToC
   @media (min-width: 768px) {
     .\@xl\:md\:max-w-\[25\%\] {
       max-width: 25%;
-    }
-
-    .\@xl\:md\:max-w-\[35\%\] {
-      max-width: 35%;
-    }
-
-    .\@xl\:md\:max-w-\[40\%\] {
-      max-width: 40%;
-    }
-
-    .\@xl\:md\:max-w-\[30\%\] {
-      max-width: 30%;
     }
 
     .\@xl\:md\:text-lg {
