+++ CHANGELOG.mdx
@@ -16,6 +16,10 @@ Prefix the change with one of these keywords:
 
 ## [Unreleased]
 
+### Changed
+
+- Examples and projects
+
 ## [0.5.6]
 
 ### Fixed
