+++ src/examples/blocks/EventCards.stories.tsx
@@ -5,10 +5,10 @@ import { Main } from '../../layouts/Main/Main'
 import { FooterBasic } from '../../components/Footer/FooterBasic/FooterBasic'
 import { Column } from '../../layouts/Column/Column'
 import { TopBar } from '../../components/TopBar/TopBar'
-import { HeroTextImage } from '../../_deprecated/HeroTextImage/HeroTextImage'
 import { Heading } from '../../components/Heading/Heading'
 import { Card } from '../../components/Card/Card'
 import { EventData } from '../../data/EventData'
+import { TextImage } from '../../components/TextImage/TextImage'
 
 const meta: Meta = {
   title: 'Examples/Blocks',
@@ -46,9 +46,15 @@ export const EventCardOptions: Story = {
 
       <Main>
         <Section hasProse>
-          <HeroTextImage maxWidth="5xl" hasBorder>
-            <HeroTextImage.Content title="Event Card Options" headerType="h1" />
-          </HeroTextImage>
+          <TextImage hasBorder>
+            <TextImage.Content headerType="h1" title="Event Card Options">
+              <p>
+                Nobis voluptatem dolorum et eum doloremque cupiditate velit. Praesentium architecto a distinctio aut
+                reprehenderit ducimus. Perferendis excepturi delectus nihil voluptatem non. Molestiae quas dolores
+                accusamus in. Praesent quis ligula quis nulla malesuada tempor.
+              </p>
+            </TextImage.Content>
+          </TextImage>
 
           <Container>
             <Heading text="Event Cards: Default" />
