+++ src/examples/blocks/_template.tsx
@@ -1,0 +1,77 @@
+import type { Meta, StoryObj } from '@storybook/react'
+import { Section } from '../../layouts/Section/Section'
+import { Container } from '../../layouts/Container/Container'
+import { Main } from '../../layouts/Main/Main'
+import { FooterBasic } from '../../components/Footer/FooterBasic/FooterBasic'
+import { TopBar } from '../../components/TopBar/TopBar'
+import { Heading } from '../../components/Heading/Heading'
+import { TextImage } from '../../components/TextImage/TextImage'
+
+const meta: Meta = {
+  title: 'Examples/Blocks',
+}
+
+export default meta
+type Story = StoryObj
+
+export const EventCardOptions: Story = {
+  render: () => (
+    <>
+      <TopBar>
+        <TopBar.Primary>
+          <TopBar.Logo link="https://carleton.ca/webservices">
+            <a href="https://carleton.ca" className="cu-topbar--logo">
+              <img
+                className="culogo"
+                src="https://cu-production.s3.amazonaws.com/rds/assets/cu-logos/cu-logo-color-right-horiztonal.svg"
+                width="130"
+                height="35"
+                alt="Logo"
+              />
+              <img
+                className="cushield"
+                src="https://cu-production.s3.amazonaws.com/rds/assets/cu-logos/cu-shield-color.svg"
+                width="28"
+                height="35"
+                alt="Logo"
+              />
+            </a>
+          </TopBar.Logo>
+          <TopBar.Aside />
+        </TopBar.Primary>
+      </TopBar>
+
+      <Main>
+        <Section hasProse>
+          <TextImage>
+            <TextImage.Content headerType="h1" title="Template">
+              <p>
+                Nobis voluptatem dolorum et eum doloremque cupiditate velit. Praesentium architecto a distinctio aut
+                reprehenderit ducimus. Perferendis excepturi delectus nihil voluptatem non. Molestiae quas dolores
+                accusamus in. Praesent quis ligula quis nulla malesuada tempor.
+              </p>
+            </TextImage.Content>
+          </TextImage>
+
+          <Container>
+            <Heading text="Heading" />
+          </Container>
+
+          <Container>
+            <Heading text="Heading" />
+          </Container>
+
+          <Container isGrey>
+            <Heading text="Heading" />
+          </Container>
+
+          <Container isGrey>
+            <Heading text="Heading" />
+          </Container>
+        </Section>
+      </Main>
+
+      <FooterBasic />
+    </>
+  ),
+}
