--- src/examples/projects/cutheme/PeopleProfile.stories.tsx
@@ -1,124 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react'
-
-import { Main } from '../../../layouts/Main/Main'
-import { Section } from '../../../layouts/Section/Section'
-
-import { TopBar } from '../../../components/TopBar/TopBar'
-import { FooterBasic } from '../../../components/Footer/FooterBasic/FooterBasic'
-import { TopBarDataSingle } from '../../../data/TopBarData'
-import { TextImage } from '../../../components/TextImage/TextImage'
-
-const meta: Meta = {
-  title: 'Examples/Projects/cutheme',
-}
-
-export default meta
-type Story = StoryObj
-
-const DoublePara = () => {
-  return (
-    <>
-      <p>
-        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aenean sit amet tortor pellentesque, posuere tellus
-        vitae, sagittis justo. Vivamus imperdiet turpis nec elit ultricies, sed tempus diam dignissim. Suspendisse
-        condimentum magna vel orci vulputate, eget vulputate neque porttitor. Suspendisse euismod, urna et gravida
-        volutpat, tortor risus vehicula nisl, in vulputate lectus dolor viverra est. Etiam quis interdum nisi, et
-        malesuada lectus. Aliquam luctus, velit eget suscipit tincidunt, sem ex tempus turpis, quis pulvinar metus
-        sapien in urna.
-      </p>
-      <p>
-        Nobis voluptatem dolorum et eum doloremque cupiditate velit. Praesentium architecto a distinctio aut
-        reprehenderit ducimus. Perferendis excepturi delectus nihil voluptatem non. Molestiae quas dolores accusamus in.
-        Praesent quis ligula quis nulla malesuada tempor.
-      </p>
-    </>
-  )
-}
-
-export const PeopleProfile: Story = {
-  render: () => (
-    <>
-      <TopBar>
-        <TopBar.Primary>
-          <TopBar.Logo title="Raven Design System" link="https://carleton.ca/webservices">
-            <a href="https://carleton.ca" className="cu-topbar--logo">
-              <img
-                className="culogo"
-                src="https://cu-production.s3.amazonaws.com/rds/assets/cu-logos/cu-logo-color-right-horiztonal.svg"
-                width="130"
-                height="35"
-                alt="Logo"
-              />
-              <img
-                className="cushield"
-                src="https://cu-production.s3.amazonaws.com/rds/assets/cu-logos/cu-shield-color.svg"
-                width="28"
-                height="35"
-                alt="Logo"
-              />
-            </a>
-          </TopBar.Logo>
-          <TopBar.Aside />
-        </TopBar.Primary>
-        <TopBar.Secondary>
-          <TopBar.Menu>
-            {TopBarDataSingle.map((topBarLink, i) => (
-              <>
-                {!topBarLink.submenu && (
-                  <a key={i} href={topBarLink.href} className="cu-topbar--parent-link">
-                    {topBarLink.title}
-                  </a>
-                )}
-                {topBarLink.submenu && (
-                  <TopBar.SubMenu label={topBarLink.title}>
-                    {topBarLink.submenu.map((topBarSubLink, s) => (
-                      <a key={s} href={topBarSubLink.href} className="cu-topbar--more-nav">
-                        {topBarSubLink.title}
-                      </a>
-                    ))}
-                  </TopBar.SubMenu>
-                )}
-              </>
-            ))}
-          </TopBar.Menu>
-        </TopBar.Secondary>
-      </TopBar>
-
-      <Main>
-        <Section hasProse>
-          <TextImage flexRow="md" flipYsm>
-            <TextImage.People
-              designation="Mr. "
-              pronoun="he/him"
-              name="Greely Wrinkle"
-              jobTitle="Jedi Master & Jedi Council Member"
-              degrees="Certificate, Padawan Academy of Jedi Arts"
-              building="Onraed"
-              room="Sector A29"
-              email="greely@thejediyouseek.com"
-              phone="613-520-2600"
-              phoneExt="1234"
-              resume="#"
-              linkedin="#"
-              twitter="#"
-              facebook="#"
-            >
-              <img
-                src="https://fastly.picsum.photos/id/1062/400/400.jpg?hmac=zaTGri35k94fGnPFBesQ7tRVfjy6BUCtXDFQdWQ3r-k"
-                alt="People profile graphic"
-                width="350"
-                height="350"
-              />
-            </TextImage.People>
-          </TextImage>
-
-          <DoublePara />
-          <DoublePara />
-          <DoublePara />
-        </Section>
-      </Main>
-
-      <FooterBasic />
-    </>
-  ),
-}
